+++ httpClient/projections/getState.js
@@ -1,6 +1,7 @@
 var debug = require('debug')('geteventstore:getProjectionState'),
     assert = require('assert'),
     req = require('request-promise'),
+    _ = require('lodash'),
     url = require('url'),
     q = require('q');
 
@@ -13,22 +14,25 @@ module.exports = function(config) {
         return url.format(urlObj);
     };
 
-    return function(name) {
+    return function(name, options) {
         return q().then(function() {
             assert(name, baseErr + 'Name not provided');
 
-            var options = {
+            var qs = {};
+            options = options || {};
+      
+            if (!_.isEmpty(options.partition)) qs.partition = options.partition;
+            var urlOptions = {
                 uri: buildUrl(name),
                 headers: {
                     "Content-Type": "application/vnd.eventstore.events+json"
                 },
                 method: 'GET',
-                json: true
+                json: true,
+                qs: qs
             };
 
-            return req(options).then(function(response) {
-                return response;
-            });
+            return req(urlOptions);
         });
     };
 };