+++ tests/support/testPartionedProjection.js
@@ -1,0 +1,12 @@
+fromAll()
+	.foreachStream()
+	.when({
+		$init: function() {
+			return {
+				data: {}
+			};
+		},
+		TestProjectionEventType: function(state, ev) {
+			state.data = ev.data;
+		}
+	});