+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 1.1.9 (2016-06-28)
+
+#### TCP Client
+
+- Feature: add subscribeToStreamFrom
+
+#### HTTP Client
+
+- Feature: get state of partitioned projection
+
+#### Dependencies
+
+- replace underscore with lodash
+- upgrade version event-store-client 0.0.7
+
 # 1.1.8 (2016-06-20)
 
 #### HTTP Client